+++ src/app/pages/demo-components/filter-components/demo-filter-range-slider/demo-filter-range-slider.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { DemoFilterFilterRangeSliderComponent } from './demo-filter-range-slider.component';
+
+describe('DemoFilterFilterRangeSliderComponent', () => {
+  let component: DemoFilterFilterRangeSliderComponent;
+  let fixture: ComponentFixture<DemoFilterFilterRangeSliderComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [ DemoFilterFilterRangeSliderComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(DemoFilterFilterRangeSliderComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});