+++ src/app/pages/demo-components/filter-components/reset-all/reset-all.component.ts
@@ -1,6 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { MdsFilterResetComponent, MdsHightlightPrismModule } from 'medes-ui';
+import { SelectedFilterInterface, SelectedFilterSliderInterface } from 'medes-ui/lib/mds-filter/mds-filter.interface';
 
 @Component({
   selector: 'mds-reset-all',
@@ -10,10 +11,20 @@ import { MdsFilterResetComponent, MdsHightlightPrismModule } from 'medes-ui';
   imports: [CommonModule, MdsHightlightPrismModule, MdsFilterResetComponent]
 })
 export class ResetAllComponent {
-  selected = { category: ['shoes'] };
+  filterSelected: SelectedFilterInterface = {};
+  filterRangeSelected: SelectedFilterSliderInterface = {
+    min: 0,
+    max: 2500,
+    start: 100,
+    end: 2000
+  }
 
 componentdemo = `
 <!-- Reset Filter Component -->
-<mds-filter-reset [class]="'btn btn-primary'" [content]="'Reset All'" [(filterSelected)]="selected"></mds-filter-reset>`;
+<mds-filter-reset [class]="'btn btn-primary'"
+  [content]="'Reset All'"
+  [(filterSelected)]="filterSelected"
+  [(filterRangeSelected)]="filterRangeSelected"
+></mds-filter-reset>`;
 
 }