+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.14.3-beta.1 - (Next Release)
+- New Component - Filter Range Slider
+- Component Enhancement - Reset Filter able to reset new filter range slider
+- Component Enhancement - Add error message if required value not setted
+- New Pipe - filter data base on min & max value
 ### Takedown & Republish library (Republish - 231122)
 - All medes-ui libraries version major version will be renamed put to minor
 - For example 13.0.1 will be changed to 1.13.1, if example version 14.0.2 will be 1.14.2